auth/requirements.txt
backoff==2.2.1
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
flask-swagger-ui==4.11.1
gevent==22.10.2
gunicorn==20.1.0
oauthlib==3.2.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pydantic==1.10.4
redis==4.5.1
SQLAlchemy==2.0.2
flask-sqlalchemy==3.0.3
email-validator==1.3.1
requests-oauthlib==1.3.1
opentelemetry-api==1.16.0
opentelemetry-sdk==1.16.0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-exporter-jaeger==1.16.0
jaeger-client==4.8.0
Flask-OpenTracing==1.1.0
redis-rate-limiter==0.2.0
user_agents==2.2.0
sentry-sdk==1.19.0
blinker==1.6